Factor out common delayed message dispatching
