+++ config/services_integration_sync.yaml
@@ -13,3 +13,9 @@ services:
         arguments:
             $recheckPeriodInSeconds: '%job_timeout_check_period%'
             $enabled: false
+
+
+    App\MessageDispatcher\JobCompletedCheckMessageDispatcher:
+        arguments:
+            $recheckPeriodInSeconds: '%job_completed_check_period%'
+            $enabled: false
