+++ src/MessageDispatcher/TimeoutCheckMessageDispatcher.php
@@ -7,21 +7,9 @@ namespace App\MessageDispatcher;
 use App\Event\JobReadyEvent;
 use App\Message\TimeoutCheckMessage;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
-use Symfony\Component\Messenger\Envelope;
-use Symfony\Component\Messenger\MessageBusInterface;
-use Symfony\Component\Messenger\Stamp\DelayStamp;
 
-class TimeoutCheckMessageDispatcher implements EventSubscriberInterface
+class TimeoutCheckMessageDispatcher extends AbstractDelayedMessageDispatcher implements EventSubscriberInterface
 {
-    public const MILLISECONDS_PER_SECOND = 1000;
-
-    public function __construct(
-        private MessageBusInterface $messageBus,
-        private int $recheckPeriodInSeconds,
-        private bool $enabled = true
-    ) {
-    }
-
     public static function getSubscribedEvents()
     {
         return [
@@ -31,15 +19,8 @@ class TimeoutCheckMessageDispatcher implements EventSubscriberInterface
         ];
     }
 
-    public function dispatch(): void
+    protected function createMessage(): object
     {
-        $message = new TimeoutCheckMessage();
-        $envelope = new Envelope($message, [
-            new DelayStamp($this->recheckPeriodInSeconds * self::MILLISECONDS_PER_SECOND)
-        ]);
-
-        if ($this->enabled) {
-            $this->messageBus->dispatch($envelope);
-        }
+        return new TimeoutCheckMessage();
     }
 }
